src/test/java/com/caio/produto_api/service/ProdutoServiceTest.java
package com.caio.produto_api.service;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import java.util.Arrays;

import org.junit.jupiter.api.BeforeEach;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.caio.produto_api.model.Produto;
import com.caio.produto_api.repository.ProdutoRepository;

public class ProdutoServiceTest {

    @InjectMocks
    private ProdutoService service;

    @Mock
    private ProdutoRepository repository;

    @BeforeEach
    public void setup(){
        MockitoAnnotations.openMocks(this);
    }

    public void listarTodos_deveRetornarListaProdutos(){
        var produtos = Arrays.asList(new Produto(), new Produto());
        when(repository.findAll()).thenReturn(produtos);

        var resultado = service.listarTodos();

        assertEquals(2, resultado.size());
        verify(repository, times(1)).findAll();
    }
    
}
